refactor(owner-view): converge loan/investment/asset onto public layout (SSOT)
Slice 3, same pattern as #341. These 3 already had rich PublicEntityDetailPage
configs inline in their PUBLIC routes; only their dashboard routes were the flat
EntityDetailPage. Now both routes share one config and owners get the public layout
+ manage bar.

- New src/components/public/detail-configs/{loan,investment,asset}.tsx — configs
  moved verbatim with helpers/imports (loan: calculateProgress, PublicEntityCTA,
  EntityData typing, mobileStickyCTA, renderSidebarExtra, showPaymentSection:false;
  investment: invFmt, INVESTMENT_* maps, FundingProgress; asset: ASSET_* maps,
  displayBTC, BookEntityButton, showPaymentSection:false). Zero behavior change.
- index.ts registry extended: DETAIL_CONFIGS now covers product, service, cause,
  research, event, ai_assistant, loan, investment, asset.
- Public routes slimmed to import the shared config; dashboard routes
  (/dashboard/{loans,investments,assets}/[id]) now render PublicEntityDetailPage.

Remaining flat-everywhere types: circles, wishlists, groups, documents (truly net-new
configs) — the last chunk.

tsc 0; lint clean.
